requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
python-multipart==0.0.6
pdfminer.six==20221105
python-docx==1.1.0
SQLAlchemy==2.0.35
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0

